+++ Models/RevLogMetadata.cs
@@ -5,11 +5,11 @@
         public double id { get; set; }
         public double cid { get; set; }
         public double usn { get; set; }
-        public double ease { get; set; }
+        public int ease { get; set; }
         public double ivl { get; set; }
         public double lastIvl { get; set; }
         public double factor { get; set; }
         public double time { get; set; }
-        public double type { get; set; }
+        public int type { get; set; }
     }
 }
